+++ Sources/Stock.swift
@@ -25,8 +25,8 @@ class Stock: PostgresSqlORM, JSONConvertible {
         stockId = this.data["stockid"] as? Int ?? 0
         storeId = this.data["storeid"] as? Int ?? 0
         articleId = this.data["articleid"] as? Int ?? 0
-        stockQuantity = this.data["stockquantity"] as? Double ?? 0
-        stockBooked = this.data["stockbooked"] as? Double ?? 0
+        stockQuantity = Double(this.data["stockquantity"] as? Float ?? 0)
+        stockBooked = Double(this.data["stockbooked"] as? Float ?? 0)
         stockCreated = this.data["stockcreated"] as? Int ?? 0
         stockUpdated = this.data["stockupdated"] as? Int ?? 0
     }
